backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
firebase-admin==6.8.0
pydantic-settings==2.10.1
email-validator==2.2.0
python-dotenv==1.1.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
